Toggle fixes
